requirements.txt
# Core Libraries
networkx>=3.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
rich>=13.0.0
fastapi>=0.109.0

# Visualization
matplotlib>=3.7.0
pyvis>=0.3.0
graphviz>=0.20.0
scikit-learn>=1.3.0
plotly>=5.18.0

# UI
streamlit>=1.32.0
streamlit-option-menu==0.3.6

# NLP & Embeddings
torch>=2.0.0
protobuf>=4.21.0
sentence-transformers>=2.2.0
nltk>=3.8.1
spacy>=3.7.0

# Graph Workflow & Orchestration
langgraph>=0.0.25
langchain-core>=0.1.0
langsmith>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.10
ollama>=0.1.6
langchain-ollama>=0.1.0

# File Monitoring
watchdog>=3.0.0

# PDF Parsing
pypdf>=5.4.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.0

# Additional dependencies
anyio==4.2.0
attrs==23.2.0
blinker>=1.7.0,<2.0.0
cachetools==5.3.2
certifi==2024.2.2
click==8.1.7
decorator==5.1.1
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
huggingface-hub==0.20.3
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpickle==3.0.3
jsonschema==4.21.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mpmath==1.3.0
packaging>=20.9,<24.0
pillow==10.2.0
prompt_toolkit==3.0.43
requests>=2.31.0
urllib3==2.2.0
tornado==6.4
typing-extensions==4.9.0

# Development dependencies
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
pylint>=3.0.0

# Install the package in editable mode with all dependencies
-e .

python-multipart>=0.0.6
